third_party/blink/web_tests/external/wpt/editing/run/createlink-with-selecting-img-expected.txt

This is a testharness.js-based test.
[FAIL] createLink should wrap the <img style="display:block>"
  assert_equals: expected "<a href=\\"foo.html\\"><img src=\\"/img/lion.svg\\" style=\\"display:block\\"></a>" but got "<img src=\\"/img/lion.svg\\" style=\\"display:block\\">"
Harness: the test ran to completion.
